Used the passport methods as a service.
