+++ package.json
@@ -20,8 +20,7 @@
     "mongoose": "^5.10.15",
     "morgan": "~1.9.1",
     "passport": "^0.4.1",
-    "passport-local": "^1.0.0",
-    "uuid": "^8.3.1"
+    "passport-local": "^1.0.0"
   },
   "devDependencies": {
     "jest": "^26.6.3"
